read manifest from url or fs
